refactor: improve cli and terminal experience
This is a refactor to move the logging towards utilzing the `log` macros
instead of direct `println!` calls and generally make the experience
better and more in line with other tools.
It removes the clearing of the screen in rapid-web which makes
it much easier to use together with `next dev` for example (#56).
It changes the logger from env_logger to pretty_env_logger in both
rapid-web and rapid-cli, and only uses `try_init` in the rapid-web crate
so that it can be overriden by the end user, simply by initalizing their
own logger first.
It removes the [rapid-web::logger] output since the logging crate
can be adjusted to handle that information, if neccesarry.
It also removes the `>>> R A P I D` output, to be more in line with how
minimal other tools in the JS ecosystem are.
Removes `std::thread::sleep()` calls, as they just slow the process
down without adding anything to the experience.

This PR on a more unrelated note also changes the rapid-web import in
the examples to use local paths rather than pulling from crates.io.

Closes #56